requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
yfinance==0.2.51
pandas==2.2.3
ta
APScheduler==3.10.4
pytz==2024.2
python-dotenv==1.0.1
